+++ src/main/java/br/com/zalf/prolog/webservice/integracao/avacorpavilan/AvaCorpAvilanConverter.java
@@ -33,7 +33,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-import static br.com.zalf.prolog.webservice.integracao.avacorpavilan.AvaCorpAvilanUtils.createDatePattern;
+import static br.com.zalf.prolog.webservice.integracao.avacorpavilan.AvaCorpAvilanUtils.createDateTimePattern;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 
@@ -115,16 +115,17 @@ public final class AvaCorpAvilanConverter {
     }
 
     @VisibleForTesting
-    public static IncluirMedida2 convert(@NotNull final Afericao afericao) {
+    public static IncluirMedida2 convert(@NotNull final Afericao afericao) throws ParseException {
         checkNotNull(afericao, "afericao não pode ser null!");
 
         final IncluirMedida2 incluirMedida2 = new IncluirMedida2();
 
-        // seta valores
+        // Seta valores.
+        incluirMedida2.setCpfColaborador(String.valueOf(afericao.getColaborador().getCpfAsString()));
         incluirMedida2.setVeiculo(afericao.getVeiculo().getPlaca());
         incluirMedida2.setTipoMarcador(AvaCorpAvilanTipoMarcador.HODOMETRO);
         incluirMedida2.setMarcador(Math.toIntExact(afericao.getVeiculo().getKmAtual()));
-        incluirMedida2.setDataMedida(createDatePattern(afericao.getDataHora()));
+        incluirMedida2.setDataMedida(createDateTimePattern(afericao.getDataHora()));
         // Placas carreta 1, 2 e 3 nunca serão setadas. No ProLog apenas um veículo será aferido por vez. Caso a carreta
         // seja aferida, então a placa dela será setada em .setVeiculo()
 
@@ -508,14 +509,29 @@ public final class AvaCorpAvilanConverter {
         checkNotNull(afericaoFiltro, "afericaoFiltro não pode ser null!");
         final Afericao afericao = new Afericao();
         afericao.setCodigo((long) afericaoFiltro.getCodigoAfericao());
-        // Avilan não salva hora de realização da aferição, apenas data.
-        afericao.setDataHora(AvaCorpAvilanUtils.createDatePattern(afericaoFiltro.getDataRealizacao()));
         afericao.setKmMomentoAfericao(afericaoFiltro.getOdometro());
 
-        // Avilan não salva colaborador que realizou a aferição.
+        if (afericaoFiltro.getDataRealizacao().length() > AvaCorpAvilanUtils.AVILAN_DATE_PATTERN_STRING_SIZE) {
+            // Antes da integração, não era salvo no ERP da Avilan a hora da aferição, apenas a data. Se o tamanho da
+            // String for menor ou igual ao pattern de data, então essa é uma aferição antiga que tem apenas a data
+            // setada.
+            afericao.setDataHora(AvaCorpAvilanUtils.createDateTimePattern(afericaoFiltro.getDataRealizacao()));
+        } else {
+            afericao.setDataHora(AvaCorpAvilanUtils.createDatePattern(afericaoFiltro.getDataRealizacao()));
+        }
+
         final Colaborador colaborador = new Colaborador();
-        colaborador.setNome("Colaborador não informado");
-        colaborador.setCpf(0);
+        if (afericaoFiltro.getColaborador() == null
+                || Strings.isNullOrEmpty(afericaoFiltro.getColaborador().getNome())
+                || Strings.isNullOrEmpty(afericaoFiltro.getColaborador().getCpf())) {
+            // Antes da integração, não era salvo no ERP da Avilan quem fez a aferição. Aferições antigas não terão
+            // colaborador vinculado.
+            colaborador.setNome("Colaborador não informado");
+            colaborador.setCpf(0);
+        } else {
+            colaborador.setNome(afericaoFiltro.getColaborador().getNome());
+            colaborador.setCpf(Long.parseLong(afericaoFiltro.getColaborador().getCpf()));
+        }
         afericao.setColaborador(colaborador);
 
         final Veiculo veiculo = new Veiculo();